requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.6
django-allauth==0.57.0
django-autoslug==1.9.9
django-crispy-forms==2.1
django-phone-field==1.8.1
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
